Moved Properties form functionality onto main form
Ok reasoning time:
So basically I was running out of space. I was about to add the terrain options, and with the transport stuff in the pipeline, I realised it was finally time to move away from creating a huge form. If I put it in on the main form people will know that the option to customize stuff is there, and there will be more space and ui options that I can work with (I am by no means a UX person). There is obviously the problem that my simplistic ui knowledge falls into the obvious-programmer-territory but given the amount of options and customization we are about to have I think this is the right move.

(There are a lot of bugs from this commit, this is not stable)